Run package installation test on SDK images
